prefactor: extract to useStatusCondition composable
